+++ BattleNoid_V2/Assets/Terasurware/Classes/Entity_Player.cs
@@ -13,10 +13,10 @@ public class Entity_Player : ScriptableObject
 		public int index;
 		public string code;
 		public string name;
-		public double baseHp;
-		public double baseDamage;
-		public double baseMoveSpeed;
-		public double baseRange;
+		public float baseHp;
+		public float baseDamage;
+		public float baseMoveSpeed;
+		public float baseRange;
 		public int cost;
 	}
 }
